move qless spec job to support dir
